+++ src/lbm_image.c
@@ -789,7 +789,7 @@ static int size_acc(lbm_value v, bool shared, void *acc) {
     lbm_int arr_size = lbm_heap_array_get_size(v);
     const uint8_t *d = lbm_heap_array_get_data_ro(v);
     if (arr_size > 0 && d != NULL) {
-      sa->s += 1 + 4 + arr_size;
+      sa->s += (int32_t)(1 + 4 + arr_size);
     }
   }break;
   }
@@ -1070,7 +1070,7 @@ static uint32_t last_const_heap_ix = 0;
 bool lbm_image_save_constant_heap_ix(void) {
   bool r = true; // saved or no need to save it.
   if (image_const_heap.next != last_const_heap_ix) {
-    last_const_heap_ix = image_const_heap.next;
+    last_const_heap_ix = (uint32_t)image_const_heap.next;
     r = write_u32(CONSTANT_HEAP_IX, &write_index, DOWNWARDS);
     r = r && write_u32((uint32_t)image_const_heap.next, &write_index, DOWNWARDS);
   }
@@ -1097,7 +1097,7 @@ void lbm_image_init(uint32_t* image_mem_address,
 void lbm_image_create(char *version_str) {
   write_u32(IMAGE_INITIALIZED, &write_index, DOWNWARDS);
   if (version_str) {
-    uint32_t bytes = strlen(version_str) + 1;
+    uint32_t bytes = (uint32_t)(strlen(version_str) + 1);
     uint32_t words = (bytes % 4 == 0) ? bytes / 4 : (bytes / 4) + 1;
     write_u32(VERSION_ENTRY, &write_index, DOWNWARDS);
     write_u32(words, &write_index, DOWNWARDS);
